requirements.txt
groq
openai-whisper
yt-dlp
moviepy==1.0.3
mediapipe
opencv-python
numpy
torch
tqdm
colorama
python-dotenv